+++ src/backend/core/tests/external_api/test_external_api_documents_link_configuration.py
@@ -64,7 +64,7 @@ def test_external_api_documents_link_configuration_can_be_allowed(
     resource_server_backend,
     user_specific_sub,
     mock_reset_service_connections,
-    django_capture_on_commit_callbacks,
+    capture_service_resets,
 ):
     """
     Connected users SHOULD be allowed to update the link configuration of a document
@@ -90,7 +90,7 @@ def test_external_api_documents_link_configuration_can_be_allowed(
         "link_role": models.LinkRoleChoices.EDITOR,
     }
 
-    with django_capture_on_commit_callbacks(execute=True):
+    with capture_service_resets():
         response = client.put(
             f"/external_api/v1.0/documents/{document.id!s}/link-configuration/",
             new_data,